Add explicit platform declarations for module-specific apps.
Any of the demo apps that are targeted at specific modules should
declare the platform. Without this, there are issues when creating
lockboxes.  When running torizoncore-builder to create a lockbox
the tool will attempt to pull both 32- and 64- bit images which
may not exist in all cases, notably the accelerated containers
which are architecture specific.

Signed-off-by: Drew Moseley <drew@moseleynet.net>