torizon7:weston: Create test composes
To use the weston screenshot, we need to run the container with
the debug flag.
In this demo, we have a complex command, so it's hard to edit the
container command to include this flag.

At the same time, I don't want this variable on the deploy
composes. It should work even when this variable is not defined,
but it is going to give a warning.

Signed-off-by: Allan Henrique Kamimura <allan.kamimura@toradex.com>